Changes: - Implementation of automatic photometry download for a star through Sed.py (and via the IvS.SED.builder). Download can also be done separately by calling Sed.buildPhotometry (not part of the Sed class). Sed() deals with the photometry properly and does not sort it anymore like the other data. Takes the errors along as well. PlotDust.py is not yet updated for this. ComboCode.py updated with series of new keywords, such as location of raw photometry data and removing certain photbands. When Photometry_IvS is present in Sed.dat, the photometry is automatically downloaded if not present yet.
- Replaced ComboCode.getStarGrid with getStars. Same functionality.

- The keyword POWER_T_DUST was removed. It was used for determining R_MAX of a dust 
species from T_MIN. Now, this is a fixed power law instead. It is only used for
the first iteration of an MCMax model. Second iteration, the R_MAX is taken from
the previous dust model. The use of POWER_T_DUST is not very common and pretty
pointless. Now, the dust T power law simply assumes s=1 (or power =0.40). 

- Following this, Star.powerRfromT is gone.

- Replaced some try: except ...: clauses with simple checks for model availability
to clean up coding.

- Removed Star().convertRadialUnits(). It was not used, and is too difficult
to include. Likely not necessary since the availability of giving R values in AU

- Made searching for ; and : in inputComboCode.dat easier by removing all of 
the comment ; and : characters.

- ModelManager now does NOT run GASTRoNOoM-cooling after the last MCMax iteration
if there is no mline request. (ie for dust modeling only)

- Replaced several functions throughout CC with versions included in Star() 
objects for retrieving physical information of gas or dust shell, such as the 
radial grid, the temperature, velocity, density and abundance profiles. Any 
number density requests have been consolidated in a single method getNumberDensity
with several options to specify the source GASTRoNOoM outputfile, the molecule, 
etc. This simplifies the coding significantly, with set methods for 
creating filenames from GASTRoNOoM output, and single methods for retrieving
the output. Affects: Lime.py, PlotGas.py, PlotDust.py, Star.py. 

- Determining the R_MAX of a species based on T_MIN has been simplified.

- In line with these changes, all dust temperature methods have been consolidated
into one, and the power law dust temperature has been removed from Star() objects. 

- MCMax.readVisibilities can now read visibilities as a function of both baseline
and wavelength, provided they were ray-traced. 

- MCMax.py now automatically tries to include .particle files when full scattering
is requested. If the .particle file is not found, it will still revert to the 
file pointer that is in Dust.dat. If .particle files are listed in Dust.dat, they
will still be included as such, but it is recommended to give .opacity files there

- Star() objects now setthe default SCATTYPE to 'ISOTROPIC' and T_CONTACT to 0.

- Star()['DUST_LIST'] has been replaced with Star().dust_list and does not 
function as a dict keyword anymore. This for consistency reasons and because it's
technically not a property to be calculated, but rather a fixed property of the
model based on the requested dust species. Star() objects now have a method that
returns the tuple with the included dust species: Star.Star().getDustList(). 
The star.dust_list is sorted based on Dust.dat, with the MRN settling species 
(rgrains) listed first. MCMax.py doesn't do this anymore as it is not necessary.
Affects: MCMax.py, ComboCode.py, PlotDust.py, TrendAnalysis.py

- Following the above, Star.py now handles dust species a lot more consistently.
Upon initalization dust is not yet set (so easy copying of a Star() model and
changing the requested abundances/species is still possible). The first time
the dust_list is requested through above method, the list is initialized if not
yet available through readDustProperties(). This also sets all properties from 
Dust.dat relevant for the requested species. Includes ALL properties, such as
PART_FILE or MOLAR_WEIGHT. Other modules now refer to these properties in Star.py
if they make use of a Star() object as input. Does not save info of dust species
not included in the model.
Affects: MCMax.py, ColumnDensity.py, Star.py

- Removed many calls to Dust.dat, in favour of using the Star().dust_info. This
simplifies the coding significantly, e.g. in calcSPEC_DENS_DUST()

- Removed the profilers module in cc.tools. Profiler.py and CustomOpacity.py 
have been moved to cc.modeling.tools. 

- Profiler.py now contains the DustTemperaturePowerLaw function following the 
typical form in the Rayleigh limit. It now takes a radial grid as input, leaving
open where this grid comes from (and not depending on old models).

- OpacityMerger.py is gone. The function mergeOpacity is now part of CustomOpa.py

- The power keyword of plotDustTemperature(/Species) now gives s in the power law
rather than the power itself. 

- Cleaned up PlotGas.py for profile plotting methods (velocity, abundances...), 
and added additional unit options such as 'm' and 'au'

- PlotDust now has a method to plot visibilities as a function of baseline. 
Work in progress!