Changes: - Sed.py now sorts MIDI data, does not sort Photometry (that is sorted upon building the file) and does not sort Sacha's ISO data. Sorting algorithm is now taken from numpy. All these data now also include error arrays. PACS/SPIRE do not. PlotDust.py and Sed.py know how to handle the different lengths of the data arrays
- MCMax.readVisibilities now reads as a function of both wavelength and baseline
and is independent of filename other than being 'visibility' or'basevis'. The 
columns are automatically associated with their proper wavelengths and baselines

- DataIO now has a new removeComments method that splits read lines based on
comment characters more efficiently than before. The method is now used by both
readDict and readCols. The comments are made available for parsing if requested.
readCols now also allows for the comments to be returned. Comment_chars in all 
cases can be chosen.

- readCols's keyword start_from_keyword is now more flexible and can be anywhere
in the document instead of the first word of a line. 

- PlotDust.py has been expanded with plotting visibilities as a function of 
baseline in addition to correlated fluxes. For the former, both the correlated
fluxes and the MIDI total spectrum are needed. Automatically determines requested
wavelengths, and selects all available baselines, but for now only for MIDI.

- PlotDust().plotSed now adds error bars to photometry points. Not yet for spectra!

- PlotDust.plotCorrflux is now more efficient in the amount of reading from disk
it has to do. It now reads all data and model files only once and remembers them
until finished with plotting. It will now also properly plot data or models 
regardless of the availability of either. The dimensions of the tile plot is now
flexible depending on how many baselines are plotted. This formalisms are also 
applied to plotting visibilities.

Bugs:
- SED_PHOT_REMOVE should give an empty list instead of [''] when no photbands
are to be removed. 
- The default value of SED_PHOT_PSUFFIX was changed to an actual location.
- If only one iterations is asked with GASTRoNOoM on (pointless, but possible), 
the gas_session in ModelManager doesn't have a molec_list because it is never
initiated. ModelManager now checks for requested molecules in the Star() object
instead
- Excluded coolrate***.dat files from Star().getGasRad() as it does not list 
radii
- The order of MIDI wavelength, flux and error needs to be reversed. For MIDI flux
this is now done in Sed.py, for correlated fluxes this is done in PlotDust.py