Conversion from SVN to git repository.
CHANGES:
- All modules and subfolders moved as a whole to the cc module *within* the home folder of the git repository. This fixes
the issue that files in the home folder were not part of the repository (eg ComboCode/__init__.py )
- New subfolder in the home folder: aux/ which contains all auxiliary files for ComboCode that do not have to be
maintained by users, unlike the contents of Data/ which is *not* part of the repository. This is not part of the cc
source code, of course, since it is not included in __init__.py.
- Files included in the aux subfolder are all the template input files (CC, GASTRoNOoM, MCMax), the lists of input
keywords for the different GASTRoNOoM modules, and the code specific mutable parameter lists. The info file of
inputComboCode.dat is technically not an auxiliary file, but it is included here as it definitely doesn't belong with
the source code.
- CC_Structure was merged into info_inputComboCode.dat for now. Will have to be updated later
- The svn usage pdf was removed.