Update README.md
Added an installation step for adding the repo to the PYTHON_PATH in the .bash_profile. 
Added an installation step to copy the distribution's usr folder to the user-specific folder (usr.dist/ -> usr/). The folder contains the settings managed by the user as required by ComboCode.
Added ComboCode subfolder structure.