Bugs: - When plotting abundance profiles from GASTRoNOoM output, ComboCode was incorrectly correcting for enhance_abundance_factor. This factor is already taken into account in the GASTRoNOoM output, while OPR and isotopic ratios are not. ComboCode now makes a clear distinction between the two. Molecule().getAbundFactor now only returns the latter. PlotGas.py now takes this into account properly when plotting abundance profiles.
Changes:
- Introduction of the usr/Path.dat file that includes numerous folders for modeling and data locations,
including the now flexible locations of GASTRoNOoM and MCMax. InputComboCode.dat was updated with all these
changes.
- To make use of Path.dat in CC, you can now import cc.path, which includes all folders defined in Path.dat as
well as cc.path.home/usr/aux. cc.path.home refers to the former 'path_combocode', which was taken out anywhere
in the code and does not have to be defined by the user anymore. It is now taken from the location in which the code was installed.
- Data objects are also made with this folder system in place. The same rules as before apply: No data objects
are made unless the paths are defined in cc.path. No default values are available yet, so if a given path is
not in Path.dat, the code will crash. Defaults will be added at a later time.
- The option to link model output folders to a different location is gone. The user can make these links
himself easily. The functionality was overkill.
- Made a clear distinction between CC input for line listing (simply plotting downloaded CDMS/JPL line lists)
and line selection (selecting lines to calculate with GASTRoNOoM from files, explicit definition, or a wavelength
range). Any line selection keyword now starts with LS_ instead of LL_.
- When LINE_SELECT=1, it now selects all lines for requested MOLECULE in the wavelength range appropriate for
either PACS or SPIRE. You do not have to define the wavelength ranges yourself anymore. This allows CC to
create line lists for SPIRE and PACS at the same time without extra effort.
- Removed the option to give abundance_filename keys for isolated filenames. You always have to specify the
path now.
- Moved the extinction curve input data to the aux/ folder. They now comes as part of the repository.
- Added the option to DataIO.getInputData to return a single element based on index.
- Cleaned up a lot of code, including Molecule.py and Transition.py. Removed the
Transition.makeTransitionsFromTransList method as it did not add much. Included its functionality in Star.py
and PlotGas.py where it was used.
- Cleaned up DustOpacit.py and fixed some of the pointers that have since changed. This module is becoming
somewhat obsolete and may disappear in the future.
- Added the option to provide the output location of plots produced by TrendAnalysis.py