Changes: - Replaced the old getBestVlsr method of Transition() objects with a more robust version. It now interpolates the model onto shifted data velocity grids, cutting down on calculation time and being independent of there being zeroes outside the line profile. Code readability significantly improved. The same interpolation method is now used in getLogLikelihood.
- Replaced the mfiltered option of PlotGas().plotTransitions with the fit_vlsr
option. Previous functionality was unclear and not required. Now, it offers the
choice to shift the line profiles to the best fitting vlsr, or not shift at all
by using the default vlsr. The code was also cleaned up and made significantly
more readable.

- DataIO.joinPdf now makes use of the PyPDF2 package, rather than OS dependent
external software (such as pdfunite or pdftk). Documentation will be updated to
reflect the additional dependency of ComboCode.