Changes: To alleviate a few inconsistencies with how N_QUAD is treated in ComboCode several changes were implemented:
1. To allow for gridding in ComboCode over N_QUAD, manual definitions of
TRANSITION keywords (which include their own n_quad) now have their n_quad
overridden by the model's N_QUAD keyword.
2. When no gridding over N_QUAD is requested, the n_quad in the TRANSITION
definitions are used instead of the model's N_QUAD, as before.
3. When LINE_SELECT=2, and a file gives the transitions, n_quad cannot be given
anymore in the file. It is assumed a file gives 11 entries. It can still take
either the GASTRoNOoM molecule name syntax, or the short-hand syntax (eg
006008.012016 versus 12C16O)
4. Radio db autosearch transitions, as well as line-select 1 and 2 always take
N_QUAD as input value for n_quad. However, if manual TRANSITION definitions
overlap with any of the former, the n_quad in the manual definition is taken
instead. No grid is created over the different n_quad values!

The standard inputfile has been updated with this information.

Transition.makeTransition has been cleaned up and standardised in terms of the
availability of a Star() object and the use of its N_QUAD, or the 12th value in
a manual TRANSITION definition.

Star.py and ComboCode.py now check whether automatic additions of transitions
overlap with manual definitions, and take the latter's n_quad if there is
overlap.

ComboCode.py now adds radio autosearch transitions before the star_grid is
created to avoid some of the inconsistencies from before. The data are still
added after the star grid is created.

Bugs:
- In case the offset in a TRANSITION definition (or even in a radio db key) were
set as 0.00, it was assumed to be different from 0.0 (due to the string format).
Now all TRANSITION objects are created with the correct offset, in case it is 0
and are correctly compared. 0 offset is 0 offset!