Implementation of the Reddening and SedStats modules.
Dust models are now reddened instead of data dereddened. The latter is still
possible manually. This also implies that the extinction is determined on a
model per model basis, which is distance dependent, making the module a lot more
robust and consistent.

SedStats provides chi^2 calculation for models with respect to photometry, both
with consistent photometric bands from the IvS repository, and other photometric
data. The former makes use of calculating photometry based on the response
function of the passband, while the latter simply interpolates the model
spectrum to the measured photometry wavelengths.

Changes:
- inputComboCode.dat now contains flags to include SPIRE, PACS, RADIO and SED
data in the modeling. They can be turned off regardless of the path being given
in Path.dat. ComboCode.py was updated with this as well.
- Some cleaning up of ComboCode.py import statements and how variable types are
handled.
- The old dereddening module of the Sed() object has been removed. The auxiliary
files for the reddening laws of cardelli and chiar/tielens have been removed.
- Ak from Star.dat has been removed. It is now always taken directly from the
interstellar extinction maps of Marshall, or otherwise Drimmel. Distance is
taken into account.
- Sed.py now contains helper methods to retrieve Ak, the photometric bands info
and determine the effective width of the photbands. it also contains a function
to calculate the photometry of a model based on the photometric bands. It was
cleaned up to get rid of the obsolete input of extinction model handling. It now
saves longitude and latitude, and ak from IvS repo. Distance is not given to
the object, but is instead passed to the getAk method. the getExtinctionModel
method was removed.
- SedStats and PlotDust have been updated to include photometry properly, with
the new methods in Sed.py and Reddening.py. Cleaned up PlotDust.py.
- SedStats was included in ComboCode.py
- Plotting2.py now has a generalized function to generate line types. This is
also called by PlotDust.py when creating SED plots to make sure spectra and
photometry are plotted correctly together. Moreover, additional functionality
for managing error bars of y and x values separately were added, see
documentation.
- BasicStats.calcChiSquared now gives the reduced chi^2, and allows to specify
the number of degrees of freedom.
- Documentation of Statistics.py updated and has been changed to accomodate Sed
statistics. A lot of keywords for setInstrument were removed, as they are
encompassed by **kwargs, which are passed along to the Instrument creation call.
- Molecule.dat was updated for some new/old molecules in terms of indices type.
- Star.dat Ak column was removed.

Bugs:
- DataIO.writeCols had trouble with detecting strings in array format elements
of the cols variable. the type() is types.StringType construction had to be
replaced with isinstance(**,str) which is more consistent.
- Made sure Pacs.py doesn't crash because it cant find a PACS database when
GASTRONOOM is turned off.