This build includes major changes to the way ray tracing of MCMax models is handled. This means some of the old keywords in the inputfile have been changed to be more consistent with each other. These changes are also reflected in the input of multiple methods in MCMax.py, as well as in setting defaults in Star.py.
Now inclination is an integral part of ray tracing. Its default is still 45 degrees as before (leading to the output name of spectrum45.0.dat and similar). Inclination is included when checking if a spectrum has already been ray traced. However, now these checks are also done for Images, visibilities and basevis (which now is its own keyword), and the only way to force a recalculation is by setting RT_REDO on, or by placing the output in its own folder (in which case the ray tracing is always performed anew). Additionally, removing the central source from the ray tracing now is a valid option for all model observations.

MCMax.py now has a single method rayTrace that combines all model observations into one, set by a single keyword rt_type. The modelfolder and inputfilename keys can be given explicitly, or can be recreated from the model id and path_mcmax. Following these changes the rayTrace call in the MCMax object has been rewritten as well.

PlotDust and SedStats had to be updated for the new ray tracing settings, as well as the inclination inclusion, for all model observation types.

Changes:
- Added a template object called MlineReader which is intended to function as SphinxReader. Not yet implemented. The Reader parent objects now allows not to replace_spaces, which is needed for reading ml3 output.
- Introduced the Meixner option for denstype. The Meixner model comes with different parameters, which are not set by default in Star.py. Beware of thise.
- Following this, the assumption that CC uses 1D spherical symmetry is not valid anymore. All methods will be updated to accommodate this in the future, but for now a getDustTheta for the theta grid has been added to the Star() object, as well as the option not to reduce the rad/theta arrays for T and density.
- In addition, the M_DUST parameter is now also calculated for denstypes other than POW, but this time from the MCMax output if available, and otherwise from the GASTRoNOoM output if available, taking into account the dust-to-gas ratio and velocity grids.

Bug:
- In updated ContinuumDivision for the new ray tracing aspects, a bug was discovered where indices may not coincide between different methods in case one of the MCMax models had failed. This is now fixed. Every model now gets its own settings for RT_SPEC and others so grids over that parameter are now possible.